refactor(cli): split docs into api/docs leaf shape
Reorganize the `docs` command's programmatic API into the fractal "leaf"
shape. `docs()` in api/docs/docs.mjs is now a dispatcher + barrel that
routes by argument shape into one leaf per envelope:

  api/docs/list/list.mjs                -> docs.list
  api/docs/detail/detail.mjs            -> docs.detail
  api/docs/detail/section/section.mjs   -> docs.detail.section

Discovery, --dense/--zh overlay loading, and topic resolution shared by
the detail and section leaves live in api/docs/_adapter.mjs; token-ref
resolution (used only by detail) stays with the detail leaf.

Pure reorganization: the `docs` export is unchanged, so api/index.mjs
and the CLI consumer (cli/commands/docs.mjs) are untouched. All --json
and human output is byte-identical; types stay central in types/docs.d.ts.
